requirements.txt
asgiref==3.7.2
celery==5.2.7
Django==4.2.1
django-cors-headers==4.0.0
djangorestframework==3.14.0
djangorestframework-camel-case==1.4.2
drf-yasg==1.21.5
httpx==0.24.1
psycopg2-binary==2.9.6
pytz==2023.3
redis==4.5.5
sqlparse==0.4.4
typing_extensions==4.6.2
coreapi